repotofiles2/requirements.txt


```
python-dotenv
aiohttp
asyncio
pydantic
langchain
pinecone-client
huggingface-hub
octoai
```